Added

+++ src/main/java/kodlamaio/hrms/business/abstracts/CandidatesService.java
@@ -1,7 +1,7 @@
 package kodlamaio.hrms.business.abstracts;
 
-import kodlamaio.hrms.business.requests.CandidatesRequest.CreatCandidateRequest;
-import kodlamaio.hrms.business.requests.CandidatesRequest.UpdateCandidateRequest;
+import kodlamaio.hrms.business.requests.CandidateRequest.CreatCandidateRequest;
+import kodlamaio.hrms.business.requests.CandidateRequest.UpdateCandidateRequest;
 import kodlamaio.hrms.business.responses.CandidatesResponse.GetAllCandidatesResponse;
 import kodlamaio.hrms.entities.concrates.Candidate;
 
@@ -10,8 +10,8 @@ import java.util.List;
 public interface CandidatesService {
     List<GetAllCandidatesResponse> getAll();
     Candidate add(CreatCandidateRequest creatCandidateRequest);
-    Candidate update(int id, UpdateCandidateRequest updateCandidateRequest);
-    void delete(int id);
+    Candidate update(Long id, UpdateCandidateRequest updateCandidateRequest);
+    void delete(Long id);
