+++ src/main/java/kodlamaio/hrms/business/responses/CandidatesResponse/GetAllCandidatesResponse.java
@@ -1,19 +1,16 @@
 package kodlamaio.hrms.business.responses.CandidatesResponse;
 
 import kodlamaio.hrms.entities.concrates.Candidate;
-import kodlamaio.hrms.entities.concrates.Jobs;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class GetAllCandidatesResponse {
 
-    private int id;
+    private Long id;
     private String identityNumber;
     private String firstName;
     private String lastName;
