+++ src/main/java/kodlamaio/hrms/dataAccess/CandidatesRepository.java
@@ -3,5 +3,5 @@ package kodlamaio.hrms.dataAccess;
 import kodlamaio.hrms.entities.concrates.Candidate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CandidatesRepository extends JpaRepository<Candidate,Integer> {
+public interface CandidatesRepository extends JpaRepository<Candidate,Long> {
 }
